Use the interface monitor for default route lookup
