+++ core/server/internal/boxdns/monitor_unix.go
@@ -1,0 +1,50 @@
+//go:build !windows
+
+package boxdns
+
+import (
+	"fmt"
+
+	tun "github.com/sagernet/sing-tun"
+	"github.com/sagernet/sing/common/control"
+	logger2 "github.com/sagernet/sing/common/logger"
+)
+
+// Keep a default-interface monitor alive even before a profile starts. The GUI
+// asks for this while it is still building the config.
+func init() {
+	logger := logger2.NOP()
+	updMonitor, err := tun.NewNetworkUpdateMonitor(logger)
+	if err != nil {
+		fmt.Println("Could not create NetworkUpdateMonitor")
+		return
+	}
+	monitor, err := tun.NewDefaultInterfaceMonitor(updMonitor, logger, tun.DefaultInterfaceMonitorOptions{
+		InterfaceFinder: control.NewDefaultInterfaceFinder(),
+	})
+	if err != nil {
+		fmt.Println("Could not create DefaultInterfaceMonitor")
+		return
+	}
+	DnsManagerInstance = &DnsManager{Monitor: monitor}
+	monitor.RegisterCallback(DnsManagerInstance.HandleSystemDNS)
+	if err = updMonitor.Start(); err != nil {
+		fmt.Println("Could not start updMonitor")
+		return
+	}
+	if err = monitor.Start(); err != nil {
+		fmt.Println("Could not start monitor")
+		return
+	}
+}
+
+var DnsManagerInstance *DnsManager
+
+type DnsManager struct {
+	Monitor tun.DefaultInterfaceMonitor
+	lastIfc control.Interface
+}
+
+func (d *DnsManager) HandleSystemDNS(ifc *control.Interface, flag int) {
+	return
+}
